Don't double-prefix view env vars (#523)
## Description

I noticed this when printing out the environment. Query params that begin with `__airplane...` will get converted into a title of the form `AIRPLANE_...`. With the current logic, we prepend _another_ `AIRPLANE_` prefix, so that the resulting env var ends up being something like `AIRPLANE_AIRPLANE_HOST`. Ideally, we have a convention that internal query params begin with `__` but not `__airplane`, but for now add some logic to prepend `AIRPLANE_` iff the env var title does not already have it.

## Test plan
Verified that environment variables were named correctly.