Views read env vars from airplane.yaml (#515)
## Description
Read views env vars from airplane.yaml config file. 

You can specify:
```
view:
  envVars:
    foo: bar
```
And we will send env var `foo: bar` to the server attached to the view.

The next step in this project is also reading env vars from the view inline config. This requires a change to the js sdk. In this PR, I'm supporting reading env vars from the inline config as soon as we change the sdk!

## Test plan
1) Added unit tests
2) I manually tested this. I was able to specify an env var in an `airplane.yaml` and send it up to be persisted in the view.