Add reverse proxy for Vite servers (#912)
## Description

If Studio is being run in a sandbox, the URL that `/dev/startView` returns is not useful, as it is a localhost address, but the view is not running locally on the user's machine. Instead, we implement a port proxy that will proxy all requests to `/dev/views/{port}/*` to `localhost:{port}` on the sandbox, and forward the responses back to our web app. This means that e.g. instead of iframing `localhost:5173`, we'll iframe `sandbox-router.fly.dev/dev/views/5173/`, which we _can_ due to the proxying logic above. cc @francescaleung 

Depends on https://github.com/airplanedev/lib/pull/513

## Test plan
Tested loading a view locally, verified that requests were appropriately proxied.