Support emitting metadata decorator (#534)
## Description

Support tsconfig property `emitDecoratorMetadata`. This is needed by some typescript packages like typeORM. One of our customers has asked for this.

esbuild does not support this out of the box because it doesn't support or run tsc. Context: https://github.com/evanw/esbuild/issues/257. The maintainer recommends using a plugin.

I read through the source code of the plugin we're using (https://github.com/thomaschaaf/esbuild-plugin-tsc). It only runs the "expensive" step to generate metadata if:
- There is a tsconfig.json file
- It has `emitDecoratorMetadata` set to true
So most of the time it will be a no-op :)

## Test plan
I wrote a unit test that fails before this change because it relies on introspecting generated metadata. it passes after this change and after setting `emitDecoratorMetadata` to true in the `tsconfig`