Make tsconfig diffs prettier (#943)
## Description
This change updates the message we show to users when we need to change their tsconfig. The current version is hard to understand because it shows the diff by key. The updated one uses more of a git-style diff (generated by a third-party library).

Resolves DTS-3

## Test plan
Testing completed successfully via running locally:

<img width="686" alt="Screenshot 2023-02-02 at 7 44 53 PM" src="https://user-images.githubusercontent.com/102088763/216483623-22068562-b8ac-47e4-9f35-96af073d9de3.png">