Fix inline Python discovery on Windows (#539)
## Description

In Python, `inspect.getabsfile` (which we use to get the file path where `airplane.task` was defined) calls [`os.path.normcase` on the result](https://github.com/python/cpython/blob/46416b9004b687856eaa73e5d48520cd768bbf82/Lib/inspect.py#L973). We did not normalize casing in the file paths of the discovered `_airplane.py` files, which means that the check below would fail on Windows and result in Python entities never being discovered. We fix this by normalizing the casing on the latter.

## Test plan
Tested on Windows and Mac