Cap number of lines returned by ecslogs (#944)
## Description
This change updates the `agents ecslogs` subcommand to cap the number of lines that are returned. I noticed that when an agent is having issues, lots of errors are generated so it's super slow to fetch the most recent 6 hours of logs, and the output size is overwhelming to users. The first couple hundred lines should be sufficient to diagnose the issue in these cases.

## Test plan
Testing completed successfully via running locally.