Check for localhost open ports for Vite (#947)
## Description

Previously, the `isPortOpen` check for vite ports would always return true, which meant that we'd always try to start views on the same port. This changes the check to take in an optional prefix so that we can search on `localhost` specifically for view ports (Vite binds to the localhost address).

## Test plan
Tested opening two views and verifying they returned different Vite server URLS.