Explicitly bind ngrok agent to us region (#952)
## Description

Ngrok will automatically attempt to start an agent in the region with the least latency. However, our tunnel domain (*.t.airplane.sh) is reserved for the `us` region, and until we support domain-specific regions, we bind the ngrok agent to the `us` region.

## Test plan
Verified that the `us` region was bound.