Add optional token to vite base path (#546)
## Description

Adds an optional token to the vite base path, which will be used when proxying views to vite. Vite does not support appending query params or setting headers in resource requests, and so we encode the token in the path instead. This will be set by the CLI when running in tunnel mode to properly authenticate requests to Vite.

## Test plan
Tested loading a view with a token in the base path.