Authenticate port proxy (#949)
## Description

Adds authentication when proxying views through the dev server in tunnel mode. Adds middleware logic to parse the encoded dev view URL for a token and verify that it matches the dev token that was registered when the dev server booted up.

Also sets the server host when starting the dev server with `--tunnel` so that views are appropriately proxied.

Depends on https://github.com/airplanedev/lib/pull/546

## Test plan
Verified that views loaded properly in --tunnel mode