Always include bundle dependencies in shim package.json (#550)
## Description

If a user has the `typescript` dependency in their `package.json` (or any of the other dependencies below), we previously would not include it in our shim package.json. This was problematic because for bundle builds, we require `typescript`. We make a change to always include these dependencies in the shim package.json so that our esbuild script can run properly.

## Test plan
Tested with a local bundler and verified that inline view deploys succeeded.