Make commands in shell images executable by non-root users (#551)
## Description
This change updates our bundle shell Dockerfile template to copy in files with permissions set to `755` instead of `700`. The problem with the latter is that if the image is being run with a non-root user, that user won't have permissions to read and execute the entrypoint.

A user ran into issues here when trying to build a shell task around the AWS glue base image. This image executes commands as a user called `glue_user`, not `root`.

cc: @leeweisberger 

## Test plan
Testing completed successfully via building some fake images locally and verifying that I could execute the entrypoint as a non-root user with these changes.