Automatically re-download corrupted builtins binary (#951)
## Description

Receiving `signal: killed` from a task execution is a strong indicator that the builtins binary has been corrupted. This PR makes it so that we automatically detect these errors and download the binary again. It's still not entirely clear why/when these errors occur, and they've significantly decreased in frequency since we started downloading the binary when the local dev server starts up, but they do still seem to appear for some users and this will help remediate those cases.

Depends on https://github.com/airplanedev/lib/pull/547

## Test plan
Manually corrupted the builtin binary and verify that the builtin binary was automatically downloaded on next execution.