Add decompression middleware into dev servers (#963)
## Description
This change updates the dev server mux to support compressed request bodies. This is needed if any clients are using the updated client logic added in https://github.com/airplanedev/lib/pull/552.

Note that the middleware is very similar to the one added to the API in https://github.com/airplanedev/airport/pull/4069. Unfortunately, we can't move that to lib at the moment and share it between the two repos because it depends on the airport-specific errorlib (which has also been semi-replicated on the CLI side).

cc: @calufornia 

## Test plan
Testing completed via extra unit tests and running locally.