Generate alias -> resource map in headless dev (#969)
## Description

Modifies the `GenerateAliasToResourceMap` so that it can be used by headless dev without needing to pass in a state. This was panicking before when attempting to call `state.UseFallbackEnv`, as state was nil for headless dev. Also adds a test for this method.

Also fixes an issue where the default dev config file was not getting loaded in properly in headless dev.

Also makes it so that we capture all panics in the CLI, and displays the panic stacktrace only when the user has enabled the `--debug` flag.

## Test plan
Tested with `airplane dev --studio=false` and verified panics are not displayed to the user anymore unless `--debug` mode is enabled.