View yarn workspace (#556)
## Description
Support deploying views within yarn workspaces.

This took two steps:
1. Set the correct root directory during discovery. This logic should be the same as tasks.
2. Build the view with the same yarn workspace logic we have for tasks. This required a bit of a refactor.

In addition, I refactored the view build process a bit.

## Test plan
I added a unit test that tests that the view builds. Also add a unit test to test that discovery works.

To test that the view renders correctly, I deployed the view with a local version of lib. A regular view worked. A view with yarn workspaces worked as well.