Only build airplane task (#559)
## Description
Previously, we were building all js/ts files in a task project directory. I believe we were doing this because we used to not bundle with esbuild, and if you don't bundle, then esbuild does not automatically walk the dependency tree and build everything.

Since we are bundling, we can rely on esbuild to walk the dep tree and build only what is necessary. This means that unrelated files are not built during discovery.

I also changed discovery issues to be errors. It's confusing when things succeed but no tasks were discovered.

## Test plan
I added a unit test with a bad file that is unrelated from the airplane task. Previously, this failed. Now it succeeds.