Optimize local javascript task execution (#557)
## Description

Speeds up local execution of JS+TS tasks by optimizing how we build local JS tasks. Because shim dependencies and the shim itself should not change between runs (or tasks), we don't attempt to install shim dependencies or build the shim if they already exist.

However, if our build script (`esbuild.js`) or any of the build dependencies change, we do want to rebuild the shim. To detect this scenario, we track a dependency hash file `dep-hash` the detects whether or not either of these has changed since the last run, and if so we rebuild the shim. This is similar to how we force `Vite` to regenerate its dependency cache. (https://github.com/airplanedev/cli/pull/886).

Note that `npm install` does not reinstall dependencies if they're already installed, but I've found that it still requires ~1s to check. Checking the dependency hash to see if we need to reinstall is much quicker.

Also adds more debug logs to see where JS task execution might be bottle-necked. 

Also writes the built entrypoint to an individual run subdirectory instead of the task subdirectory—we were seeing some occasional race conditions where one run was causing the deletion of the task directory while another run was occurring, causing the second run to error. We now keep a task subdirectory around (similar to how we keep a view subdirectory in `.airplane-view`, and cleanup run subdirectories on the fly.

After these changes, the `.airplane/` directory looks like the following:
```
.airplane/
|_dep-hash
|_esbuild.js
|_node_modules
|_package.json
|_package-lock.json
|_my_task
   |_run1234
      |_{entrypoint}
   |_run2345
      |_{entrypoint}
```

Resolves DTS-329

## Test plan
Tested local JS execution and verified that run preparation time (after the 1st for a task across all Studio sessions) went from 3-7s to 0.4-0.8 seconds.