Don't load system env vars into studio (#979)
## Description

This is causing issues in Views when the user has too many system environment variables, and the view URL becomes too long. The user should really be including any desired environment variables in the view or task definition, or `airplane.dev.yaml`.

## Test plan
Verified that system environment variables are not loaded into tasks or views but that env vars from definitions and dev config file still work.