Fix AppCondition cache (#985)
## Description

`AppCondition` is an eager cache for errors + warnings & isn't environment-aware. This patch changes it to a lazy cache that _is_ environment-aware. When something on the filesystem changes, the whole cache is thrown out, which is erring on the side of refreshing the cache very often--but because it's now a lazy cache, I don't think this is too problematic.

This patch also removes the `resources` -> `state` dependency + the unused `UnattachedResource` struct.

Resolves DTS-379.

## Test plan

Started a studio server, navigated to a task page that used resources that existed in the fallback environment. Task page now does not indicate that resources are missing.