Remove Definition interface (#567)
## Description

Removes the definition interface. We used this when migrating to `Definition_0_3` from earlier definition formats. We have deprecated them and don't have plans to introduce additional changes to this definition format that would require the use of this interface. Removing this interface makes definitions easier to work with by removing one layer of indirection.

## Test plan

Tests should pass.