Expose custom headers for CORS (#993)
## Description

I noticed that API requests in studio were getting retried, despite `X-Airplane-Retryable: false` on the response. Turns out it's not accessible in CORS mode since we aren't exposing this header to user-land.

I also added a few other headers related to gzip-encoding. I'm not sure if I got all of them. cc @yolken-airplane 

## Test plan

Triggered a failed API request from studio and confirmed it was not retried.