Differentiate run cancellation from other killed processes (#994)
## Description

Right now, when you cancel a run, it also attempts to redownload the builtins binary because context cancellations have the same exit code as killed processes (-1). This fixes the check to not include run cancelations when checking if the builtins binary is corrupted.

## Test plan
Tested cancelation locally and verified that builtin binary was not re-downloaded.