Fix JS task execution on Windows (#543)
## Description

JS tasks do not work in Windows—we get an error saying that the shim failed to build. This is because we attempt to manually "stringify" the entrypoints argument of the esbuild script by wrapping its elements in quotes. This is error-prone and does not work on Windows, because backslashes need to be properly escaped.

Instead, we just construct the underlying object and just use on `json.Marshal` to properly escape backslashes, etc. This maintains a better contract between `esbuild.js` and callers of it. cc @leeweisberger 

## Test plan
Tested on Windows. Verified that changes do not affect POSIX.