Break the install script when there are lists of commands (#997)
## Description

In lists of commands (e.g. `foo && bar`), if the first command returns a non-zero exit code but the second one doesn't, the bash script keeps rolling even though `set -e` is specified. Adding `|| false` to each command forces the bash script to stop executing if one of the commands in the list returns a non-zero exit code.

## Test plan

Tested installing invalid dependencies in a Node task. Previously this would not signal an error in installing dependencies; now it does.