Add file size to  endpoint (#1002)
## Description
Augments the `/dev/files/list` endpoint in the dev server to also return file size. We'll use this information in the frontend to determine whether to not display a file if it's too large.

## Test plan
Verified that hitting `http://localhost:4000/dev/files/list` locally returned the file sizes for all the files.