Incrementally find vite port (#1001)
## Description

Vite sometimes fails to start on the port that's randomly allocated from `localhost:0` (there's an error about how the port is already allocated). We just match the process that we do in `views/dev.go` when no port is specified, which is to start from port 5173 and incrementally find an open port.

## Test plan
Tested locally