Update views to 2.0 (#587)
## Description
Update views to 2.0.0 everywhere.

Most of these are tests and the update should have no effect.

The one thing that does have an effect is  `pkg/build/package.json`. This only has effect if the user doesn't have @airplane/views installed which is probably never the case. But if that is the case, this will cause them to use views 2.0.0