Fix task definition command for workflow runtime (#588)
## Description
This fixes an issue where if a task was deployed as a standard runtime, then changed to workflow, the command field wasn't updated and the runbroker couldn't start up.


The workflow runtime should have an empty command. The standard runtime has the command `{node}` in the db. The patch semantics in the API was seeing the command field as nil and using the previous task revision's command (which was the standard runtime). We should be passing in an empty instantiated map so that it gets cleared out if it's a workflow instead.

## Test plan
- [x] Deploy a standard task. Update the runtime to be workflow. Confirm that the command is wrong (it's `{node}` but should be `{}`. Confirm it doesn't work.
Pull in this lib change into the bundler and redeploy the workflow. Confirm the command gets wiped correctly and the workflow runbroker starts up this time.
<img width="944" alt="Screenshot 2023-03-10 at 3 51 53 PM" src="https://user-images.githubusercontent.com/8205327/224451285-e6e0f355-8032-42ae-9c07-27fccb7c8603.png">
<img width="941" alt="Screenshot 2023-03-10 at 3 51 45 PM" src="https://user-images.githubusercontent.com/8205327/224451288-4bbbee91-8486-4853-849d-8b23104b3b44.png">
<img width="944" alt="Screenshot 2023-03-10 at 4 01 52 PM" src="https://user-images.githubusercontent.com/8205327/224451280-a9566dc4-3b58-4e01-a141-55a5e139792a.png">