Remove API calls from `definitions` package (#568)
## Description

This PR removes all API client usage from the `definitions` package. This was only used for fetching resources, which we now pass in directly. This lets the caller decide if an API call is necessary.

These changes are useful in studio, where we perform operations on definitions from within the CLI's API.

## Test plan

Tests should pass.