Add prompter interface (#1011)
## Description

Adds a `Prompter` interface so that we can mock user stdin during tests. This is the [recommended approach](https://github.com/go-survey/survey/issues/448#issuecomment-1253631737) for mocking out survey calls. We have two implementations of the `Prompter` interface: a `Surveyor` that uses the `survey` package, and `Mock` that we can pre-load responses into.

Resolves DTS-250

## Test plan
Verified that this is a no-op locally.