Fix editorconfig handling of go code (#595)
## Description
This is a small change to fix the editorconfig in the golang case. Based on testing and reading the docs (https://spec.editorconfig.org/#glob-expressions), I think you have to wrap multiple matchers in braces for them to work correctly. 

## Test plan
Testing completed via opening golang files in VSCode with the new configs and verifying that indents were correct (yay!).