Add `definitions.Update` (#572)
## Description

Adds an `Update` method to the `definitions.Definition` struct. This will be used by the CLI (via https://github.com/airplanedev/cli/pull/989) when updating a task via the new config panel.

To keep this simple, I merged it's behavior with the existing `NewDefinitionFromTask` method.

## Test plan

Confirmed I could initialize a task from an existing task, which is the main usage of `NewDefinitionFromTask`.