Remove check for shell entrypoint (#593)
## Description
If a customer provides a custom `Dockerfile.airplane`, the shell entrypoint might not exist in the current filesystem or the uploaded files. This PR removes the checks for whether the entrypoint exists in some cases when the task kind is shell.

Fixes https://linear.app/airplane/issue/DTS-327/building-errors-if-shell-entrypoint-doesnt-yet-exist

## Test plan
Ran a deploy locally which generated an entrypoint via dockerfile and made sure it was correctly built + executable