Don't allow users to override build deps (#603)
## Description
While shim deps should be overridable by the user (e.g. airplane), build deps need to be the version we want so that we can successfully build tasks.

Ensure that build deps are the correct version.

## Test plan
I updated a unit test to ensure that even if the user has installed a build dep, we still use our own version of the dep.

FIXES PDS-2326