Allow optional env vars during discovery (#563)
Part of [DTS-460](https://linear.app/airplane/issue/DTS-460/environment-variables-dont-work-in-inline-config-in-studio)

## Description
Allows the passing of env vars during discovery. The CLI will use this to pass env vars to the parser during discovery for local dev.

## Test plan
I tested this end to end on a parameter variable with a dynamic default value and it works.