Fix workflows ARM builds (#615)
## Description
This change updates the lib node builder to not delete the temporalio-core aarch64 binaries. These are needed for ARM workflow builds and runs.

The images will be slightly larger without this deletion, but it's probably better to avoid platform-related problems here.

Resolves DTS-491

## Test plan
Testing completed successfully via updating the builder and running a workflow build on Mac and verifying that everything worked as expected.