Various task definition improvements (#606)
## Description

Unwraps the `timeout` and `resources` fields into `int` and `map` values, respectively. This makes it easier to construct a `Definition` struct. The structs are no longer necessary (e.g. there used to be some default timeout behavior but that was removed).

Marks a few fields in the JSON schema as non-required. These haven't been required by our API, but for some reason were marked as required in the schema.

Tweaks the serialization of options to not use label objects when not needed. This is useful in the JS transformer (see upstream).

For more context, see comments in the PR these changes were extracted from: https://github.com/airplanedev/lib/pull/590

## Test plan

I tested:
1. I can deploy a task with resources, options, and timeouts
2. I can initialize an inline and YAML-based JS task
3. The studio can discover/dev inline JS tasks