Optimistically update task cache on update (#1018)
## Description

When we update a task via the new config panel, optimistically update the task in the in-memory cache. This change will get re-applied on the next task discovery loop, but this prevents `/i/tasks/get` calls from returning the outdated task in the meantime.

## Test plan

Tested via https://github.com/airplanedev/web/pull/2220