Split up build package, part 1 (#598)
## Description
Currently, lib tests are super slow because of all the testing done in the `build` package. In order to speed things up, we need to split this package up into subpackages (e.g., for node, python, shell, etc.) that can be sharded and whose results can be cached separately.

This change starts this process by creating a few subpackages of `build`:

1. `hooks`: Install hooks logic
1. `types`: General types and name constants
2. `utils`: Utilities for dealing with templates and directory trees
3. `versions`: Logic for handling different versions

If this goes well, I'll follow up with a change that puts the different build types into their own subpackages too.

This refactoring should be a no-op from the functionality perspective.

cc: @colinking 

## Test plan
Testing to be completed via unit tests. Once the second set of refactoring is done too, I'll test end-to-end by updating cli and airport.