Detect dynamic logic within task definitions (#607)
## Description

Adds a `CanEdit` method to the runtime interface that determines if a task can be edited. Include an implementation for JS tasks. Support for Python tasks will be included in a future version.

This interface will be used by the CLI in a future PR.

## Test plan

Unit tests pass. Will be tested e2e in a future CLI PR.