Add `/i/tasks/canUpdate` endpoint (#1027)
## Description

Adds a new CLI endpoint that can be used by the studio to detect if it can update a task. This endpoint is specific to the studio, as tasks can only be updated locally if they do not contain "computed fields".

Depends on:
- https://github.com/airplanedev/lib/pull/607
- https://github.com/airplanedev/lib/pull/608

Used by:
- https://github.com/airplanedev/web/pull/2282

## Test plan

Ran locally and confirmed that tasks with dynamic logic could not be updated.