Support yarn workspaces when developing a view (#1043)
## Description
When you run `yarn install` within a yarn workspace in the `.airplane-view` folder, it errors because the `.airplane-view` folder is not part of the workspace:
![image](https://user-images.githubusercontent.com/6363419/230422438-6b7caca6-57c5-4fa7-bd7f-c8e004b635e0.png)

Here we take the last suggestion, adding an empty `yarn.lock` to signal that this folder is a separate project.

## Test plan
I set up a yarn workspace and tested that `airplane.dev` works as intended.