Sort randomized map traversals (#1045)
## Description

When iterating over a map, the results are returned in randomized order. I looked for `for` loops and `map.Values` usage and, where necessary, added a sort call. This has caused issues where lists are reordered in the UI when the API call refreshes.

## Test plan

I confirmed that config variables in the parameter form are consistently ordered. I also confirmed a few other views continued to work -- the explorer, task execution, resource list, config list.