Move builtins binaries to common location (#1046)
## Description

Currently, `airplane dev` downloads the builtins binary to the current working directory (the directory from which `airplane` was run). This causes us to pollute the user's directories since they'll have a `.airplane` folder wherever they run `airplane`, leading to many of these folders and duplicate builtins binary.

This PR makes it so that the builtins binary is instead download to `$HOME/.airplane/builtins/`. The parent directory, `$HOME/.airplane/`, already exists since the user's config is saved in this location. The builtins binary is common among all instances of airplane dev and so it's OK to have a single, global binary. We include it in a `/builtins/` subdirectory in the rare instance that we require the user to remove the builtins file.

After this change lands, `.airplane` folders should only exist in two possible sets of locations: `$HOME` and project roots.

Resolves DTS-527

## Test plan
Tested `airplane dev` locally—verified that no `.airplane` directory was created in the directory from which `airplane dev` was run, and that builtins were instead included in `$HOME/.airplane/builtins/`. Verified that local builtin execution worked.