Fix query file lookup when updating SQL tasks (#1050)
## Description

Fixes an issue with updating SQL tasks. If there is an `airplane.yaml` file, it will define the root directory instead of the directory containing the task definition. This broke the behavior for looking up the entrypoint file, which is always relative to the task definition.

## Test plan

Updated unit tests to cover this case. Tested in studio.