Explicitly set package manager for Renovate (#1067)
## Description

Renovate is unable to [updates the fixtures](https://github.com/airplanedev/cli/pull/1042). It's producing an error that appears to be caused by it using yarn v2 instead of yarn v1.

## Test plan

I tried to apply these changes directly to that PR, but can't get Renovate to retry the artifact changes without it overwriting these changes. I'm going to merge this change and then rebase that PR to see if it helped.