Account for nil prompter in InitTask (#1064)
## Description

Moves all of `cmd/airplane/testutils` to `pkg/testutils` so it can be reused outside of `cmd`. Checks for a nil prompter in places where a prompter would be invoked via `InitTask` and either picks a sane default or returns an error. Also adds an optional WorkingDirectory to `InitTaskRequest` so invoking `InitTask` does not have to rely on the actual working directory.

Resolves DTS-593.

## Test plan

Unit tests!