Return 404 when file isn't found (#1074)
## Description

Returns an ErrNotFound when supplied a nonexistent path in the `/dev/files/get` endpoint so that the Studio frontend can more easily distinguish when the endpoint errors vs when it's not found. Also adds support for relative paths in the endpoint, which is likely what our LLM's diff output will contain.

## Test plan
Verified locally