Separate patch generation from writing (#1087)
## Description

Separates out the logic for computing a patch from a unified diff from actually overwriting the file with the new contents. `gitdiff.Apply` seems to be actually seeking in the file (this is not fixed by just doing `f.Seek(0, 0)`). This was causing unexpected behavior when actually applying patches—instead, we just overwrite the contents of a new file pointer.

## Test plan
Verified locally and in conjunction with web changes + unit tests