Fix cached builds not discovering entities (#1089)
## Description
Introduce a build arg BUILD_ID into the bundle build dockerfile. We'll set this buildarg to the build id so that all the steps that occur underneath the ARG command cannot be cached, since the build id would be unique: https://stackoverflow.com/questions/35134713/disable-cache-for-specific-run-commands

The next step will be adding this build arg in the bundler (airport) when it executes the build.

## Test plan
Builds still work properly.