Add marshal method to view definition (#1110)
## Description

This PR adds a `Marshal` method to `ViewDefinition` which aligns with what we have on the task `Definition` struct. This will be used in a follow-up PR where we add logic for updating View definitions.

This PR also makes the `name` field optional, which matches the behavior of tasks. The JS SDK will automatically convert slugs into humanized strings and therefore already doesn't require a `name` field. I tested out what happens when you deploy a YAML task without a name and noticed that we don't handle it well (e.g. in the deploy logs or in the deploy UI it shows an empty string). Since this is an issue with tasks and only applies to YAML definitions, I'm punting on that for now (see https://linear.app/airplane/issue/CAP-2533/generate-names-for-tasksviews-without-names).

## Test plan

Tests pass.