Fix omitempty behavior for concurrencyLimit/timeout (#1105)
## Description

When updating a task via studio, `concurrencyLimit` was getting set to `1` in the updated task definition. This generates errors for JS/Py tasks if the SDK hasn't been updated to latest version. This is also unnecessary because `1` is the default value. The same issue existed for `timeout`, but we had special logic in the updaters for not serializing `timeout` if the value was `3600`.

This PR fixes the omitempty behavior for both fields so that they are omitted when serialized if set to their default values. It also reworks the definition serialization test to cover these cases.

## Test plan

Unit tests pass.